split test utils and misc utils fiels into internal and public
