+++ sdk/servicebus/service-bus/test/public/utils/testUtils.ts
@@ -2,11 +2,8 @@
 // Licensed under the MIT license.
 
 import chai from "chai";
-import { MessagingError, ServiceBusReceivedMessage, ServiceBusMessage, delay } from "../../../src";
+import { ServiceBusReceivedMessage, ServiceBusMessage, delay } from "../../../src";
 import * as dotenv from "dotenv";
-import { ConnectionContext } from "../../../src/connectionContext";
-import { ReceiveOptions } from "../../../src/core/messageReceiver";
-import { StreamingReceiver } from "../../../src/core/streamingReceiver";
 dotenv.config();
 
 export class TestMessage {
@@ -200,40 +197,4 @@ export enum EntityNames {
   MANAGEMENT_RULE_2 = "management-rule-2",
   MANAGEMENT_NEW_ENTITY_1 = "management-new-entity-1",
   MANAGEMENT_NEW_ENTITY_2 = "management-new-entity-2"
-}
-
-/**
- * Create and initialize a streaming receiver using a given context and entityPath.
- *
- * Defaults to peekLock, with no auto lock renewal.
- */
-export async function createAndInitStreamingReceiverForTest(
-  context: ConnectionContext,
-  entityPath: string,
-  receiveOptions?: ReceiveOptions
-): Promise<StreamingReceiver> {
-  const streamingReceiver = new StreamingReceiver(
-    context,
-    entityPath,
-    receiveOptions ?? {
-      receiveMode: "peekLock",
-      lockRenewer: undefined
-    }
-  );
-
-  let err: Error | MessagingError | undefined;
-
-  await streamingReceiver.init({
-    useNewName: false,
-    connectionId: context.connectionId,
-    onError: (args) => {
-      err = args.error;
-    }
-  });
-
-  if (err) {
-    throw err;
-  }
-
-  return streamingReceiver;
 }
